migrate price decimal(13,2), prices history  fix

+++ database/migrations/2017_03_01_110558_products_table.php
@@ -20,7 +20,7 @@ class ProductsTable extends Migration
             $table->string('name');
             $table->text('description');
             $table->string('img_src');
-            $table->float('price');
+            $table->decimal('price', 13, 2);
             // $table->float('price_provider');
             $table->float('rating');
             $table->integer('vote_count');
